% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_statistic.R
\name{basic_satistic}
\alias{basic_satistic}
\title{count basic satistic about the distribution of generator}
\usage{
basic_satistic(gen, n = 10000)
}
\arguments{
\item{gen}{generator}

\item{n}{number of sample we use to count}
}
\value{
basic satistic about the distribution of generator
}
\description{
count basic satistic about the distribution of generator
}
\examples{
gen=trans_to_gen(runif,3,4)
basic_satistic(gen)
}
