% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{get density plot about the distribution of generator}
\usage{
density_plot(gen, n = 10000)
}
\arguments{
\item{gen}{generator}

\item{n}{number of sample we use to plot}
}
\value{
density plot
}
\description{
get density plot about the distribution of generator
}
\examples{
gen=trans_to_gen(runif,3,4)
density_plot(gen)
}
