% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iverse.R
\name{inv}
\alias{inv}
\title{use inversion method to get simulation data}
\usage{
inv(n, inv_cdf)
}
\arguments{
\item{n}{number of sample}

\item{inv_cdf}{inverse function of cdf}
}
\value{
data
}
\description{
use inversion method to get simulation data
}
\examples{
inv_cdf=function(x){-log(1-x)/10}
inv(10,inv_cdf)
}
