% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{k_percentile}
\alias{k_percentile}
\title{count percentile you need about the distribution of generator}
\usage{
k_percentile(gen, p, n = 10000)
}
\arguments{
\item{gen}{generator}

\item{p}{percentile you need}

\item{n}{number of sample we use to count}
}
\value{
percentile
}
\description{
count percentile you need about the distribution of generator
}
\examples{
gen=trans_to_gen(runif,3,4)
k_percentile(gen,c(10,20,30,40,50,60,70,80,90))
}
