% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reject.R
\name{rej}
\alias{rej}
\title{use rejection method to get simulation data}
\usage{
rej(n, pdf, a, b, k)
}
\arguments{
\item{n}{number of sample}

\item{pdf}{pdf function}

\item{a}{range of data}

\item{b}{range of data}

\item{k}{bound of pdf}
}
\value{
data
}
\description{
use rejection method to get simulation data
}
\examples{
pdf <- function(x){
if(x>0&x<1){return(x)}
if(x>=1&x<2){return(2-x)}
return(0)
}
rej(10,pdf,0,2,2)
}
