% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_data.R
\name{trans_to_gen}
\alias{trans_to_gen}
\title{transform the sample function to generator,generator means only need to input n then you can get simulation data}
\usage{
trans_to_gen(sample_function, ...)
}
\arguments{
\item{sample_function}{sample function}

\item{...}{parameter in sample function exept n}
}
\value{
a function
}
\description{
transform the sample function to generator,generator means only need to input n then you can get simulation data
}
\examples{
gen=trans_to_gen(runif,3,4)
gen(10)
}
